vminit: extract cmd/vminitd into pkg/vminit/initd
Move the runnable body of cmd/vminitd (system init, ttrpc service setup,
plugin loading, signal/shutdown loop) into a new pkg/vminit/initd package
that exposes Run(ctx) and ParseFlags. The dialBackListener and the
network flag.Value type also move into the package; the latter is
exported as initd.Networks.

cmd/vminitd/main.go is now a thin wrapper that blank-imports the
shipped plugins and calls initd.Run. Downstream consumers can build
their own thin wrapper around the same package without
needing to import nerdbox's internal/ packages, which Go's internal
visibility rule prevents.

No behaviour change: signal handling, plugin load order, ttrpc setup,
and DHCP renew/release callbacks all match the previous implementation.
